Try a new machine Configuration
Try out a new setup with a GT2 belt.

Add a quick utility script to calculate the Configuration.h values from
machine measurements.

The script asks for the distance between belt pivot points. The quickest
measurement is just from the edge of one motor to the same edge of the
other. Then you can enter the diameter of the pulleys, which is
subtracted to get the width.